+++ lib/src/theme/themes/default_theme_no_secondary_border_variant.dart
@@ -251,8 +251,8 @@ class ShadDefaultThemeNoSecondaryBorderVariant extends ShadThemeVariant {
         shadows: ShadShadows.md,
       ),
       anchor: const ShadAnchorAuto(
-        preferBelow: false,
-        verticalOffset: 24,
+        followerAnchor: Alignment.topCenter,
+        targetAnchor: Alignment.topCenter,
       ),
       duration: Animate.defaultDuration,
       reverseDuration: Duration.zero,
@@ -286,7 +286,7 @@ class ShadDefaultThemeNoSecondaryBorderVariant extends ShadThemeVariant {
           width: 1,
         ),
       ),
-      anchor: const ShadAnchorAuto(verticalOffset: 24),
+      anchor: const ShadAnchorAuto(),
     );
   }
 
@@ -354,7 +354,7 @@ class ShadDefaultThemeNoSecondaryBorderVariant extends ShadThemeVariant {
       optionsPadding: const EdgeInsets.all(4),
       showScrollToTopChevron: true,
       showScrollToBottomChevron: true,
-      anchor: const ShadAnchor(offset: Offset(1, 4)),
+      anchor: const ShadAnchorAuto(offset: Offset(1, 4)),
       searchPadding: const EdgeInsets.all(12),
     );
   }
@@ -975,7 +975,11 @@ class ShadDefaultThemeNoSecondaryBorderVariant extends ShadThemeVariant {
       radius: radius,
       padding: const EdgeInsets.all(4),
       border: ShadBorder.all(color: colorScheme.border, width: 1),
-      anchor: const ShadAnchor(offset: Offset(-8, 8)),
+      anchor: const ShadAnchorAuto(
+        offset: Offset(-8, 8),
+        followerAnchor: Alignment.bottomRight,
+        targetAnchor: Alignment.bottomLeft,
+      ),
       buttonHeight: 32,
       buttonVariant: ShadButtonVariant.ghost,
       buttonSelectedBackgroundColor: colorScheme.accent,